% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Function}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{\code{data.frame} with 3 columns v1, v2 & w}

\item{y}{integer value that should be a valid vertex in the graph}
}
\value{
A resultant \code{data.frame} with the lowest computed cost for each vertex from the starting node
}
\description{
Provides the most efficient path along all the edges of a graph
}
\examples{
dijksrta(wiki_graph, 2)

}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
