\name{euclidean}
\alias{euclidean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Euclidean Function
}
\description{
Returns the greatest common divisor of 2 integers
}
\usage{
euclidean(x,y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
is an integer value
}
  \item{y}{
  is an integer value
  }
}
\details{
The function works based on the Euclidean algorithm
}
\value{
If there is no common divisor, then the function returns the default value of 1
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
\author{
Jaskirat & Filip are students at LiU
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
