% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Sample graph for the Dijkstra function}
\format{
\code{data.frame} with 3 columns
\describe{
\item{v1}{
FROM vertex used in the graph}
\item{v2}{
TO vertex used in the graph}
\item{w}{
Used to signify the cost or weight of the edge of the graph}
}
}
\usage{
wiki_graph
}
\description{
A simple data frame for the simple testing of the Dijkstra algorithm
}
\keyword{datasets}
