% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Function}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{integer value}

\item{y}{integer value}
}
\value{
A resulting number that is the greatest common divisor of the 2 input values
}
\description{
Provides the Greatest Common Divisor for 2 integer values
}
\examples{
euclidean(10, 100)
euclidean(x=10, y=100)
result_var <- euclidean(10,100)

}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
